+++ packages/zod/src/types.ts
@@ -5,18 +5,16 @@ export type EntityName = string;
 
 export type EntityC = z.ZodTypeAny;
 
-export type Define = <C extends EntityC, TSequences extends Sequences>(
+export type Define = <
+  C extends EntityC,
+  TSequences extends Sequences,
+  TContext
+>(
   Entity: C,
-  options: DefineOptions<z.TypeOf<C>, TSequences>
+  options: DefineOptions<z.TypeOf<C>, TSequences, TContext>
 ) => void;
 
 export type Manifest = <C extends EntityC>(
   Entity: C,
   overrides?: Partial<z.TypeOf<C>>
 ) => z.TypeOf<C>;
-
-export type Persist = <C extends EntityC, TContext = unknown>(
-  Entity: C,
-  overrides?: Partial<z.TypeOf<C>>,
-  context?: TContext
-) => Promise<z.TypeOf<C>>;
